+++ FrontEnd/src/api/customer.js
@@ -1,8 +1,0 @@
-import request from '@/utils/request'
-
-export function getCustomerList() {
-  return request({
-    url: '/customer/list',
-    method: 'get'
-  })
-}