+++ FrontEnd/src/api/supplier.js
@@ -1,8 +1,0 @@
-import request from '@/utils/request'
-
-export function getSupplierList() {
-  return request({
-    url: '/supplier/list',
-    method: 'get'
-  })
-}